+++ Client/package.json
@@ -16,6 +16,7 @@
   "devDependencies": {
     "@floating-ui/dom": "^1.4.3",
     "@playwright/test": "^1.35.1",
+    "@svelte-put/clickoutside": "^3.0.0",
     "@sveltejs/adapter-auto": "^2.1.0",
     "@sveltejs/adapter-node": "^1.3.0",
     "@sveltejs/kit": "^1.21.0",
