+++ Client/src/lib/ui/user/guide/panels/settings/LogOutPanel.svelte
