+++ Client/src/lib/ui/user/guide/panels/settings/index.ts
@@ -1,0 +1,2 @@
+export { default as SettingsPanel } from "./SettingsPanel.svelte";
+export { default as LogOutPanel } from "./LogOutPanel.svelte";
