+++ Client/src/lib/util/functions/click-outside.ts
@@ -1,15 +1,15 @@
 export function clickOutside(node: any) {
 	const handleClick = (event: any) => {
 		if (!node.contains(event.target)) {
-			node.dispatchEvent(new CustomEvent('outclick'));
+			node.dispatchEvent(new CustomEvent("outclick"));
 		}
-	}
+	};
 
-	document.addEventListener('click', handleClick, true);
+	document.addEventListener("click", handleClick, true);
 
 	return {
 		destroy() {
-			document.removeEventListener('click', handleClick, true);
+			document.removeEventListener("click", handleClick, true);
 		}
 	};
 }
