--- Client/src/routes/messages/+page.svelte
@@ -1,13 +1,0 @@
-<svelte:head>
-	<title>Messages &mdash; Offprint</title>
-</svelte:head>
-
-<div class="empty">
-	<h3>Pardon the dust—</h3>
-	<p>
-		This feature is still under construction.<br />Keep an eye on our
-		<a href="https://github.com/OffprintStudios/Sailfish/releases" target="_blank"
-			>GitHub Releases</a
-		> for info on when this will become available!
-	</p>
-</div>
