--- Client/src/routes/settings/about/+page.svelte
@@ -1,8 +1,0 @@
-<div class="settings-container">
-	<div class="settings-nav">hi</div>
-	<div class="settings-content">hi</div>
-</div>
-
-<style lang="scss">
-	@import "../Settings";
-</style>
